Add files via upload
This data set if from M-kopa. A company that deals with energy and Utility and give different products that they offer on Pay as you go. The company gives different types of products such as phones.The main focus is on the different types of phones that they issue like Samsung and Nokia.